ADD generiranje nakupovalnega seznama, posodobljen README.md

+++ projekt_koda/src/main/java/com/example/repository/ReceptRepository.java
@@ -1,11 +1,12 @@
-
-
 package com.example.repository;
 
 import com.example.model.Recept;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface ReceptRepository extends JpaRepository<Recept, Long> {
+public interface ReceptRepository extends JpaRepository<Recept, Integer> {
+
 }
